[refactor] API 요청 타입 분리
- api-service.ts 파일에서 타입 정의를 분리하여 request-type.ts 파일로 이동
